app/requirements.txt
flask
azure-storage-blob
psycopg2-binary
